==> src/data-prep/00_spatial.Rmd <==
---
title: "Spatial setup"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE
)
```

```{r libraries, include=F}
# Load libraries
library(tidyverse)
library(sf)
library(tigris)
library(stringr)

# Load source
source(file.path(here::here(), "src", "common.R"))

# Load shp 
ci_shp <- st_read(file.path(project_data_path, "raw", "data_EPSG_4326", "channel_islands.shp"))

# Load US counties shp 
us_counties <- tigris::counties()

## Central CA counties 
ca_counties <- us_counties %>% 
  filter(GEOID %in% c('06083', '06111', '06037'))

## MPA Atlas 
mpa_atlas <- st_read(file.path(emlab_data_dir, "mpa-atlas", "mpatlas_20201223_clean", "mpatlas_20201223_clean.shp"))
```

## Channel Islands 

Setup spatial files for the project. The shapefile for the Channel Islands was created by California Department of Fish and Wildlife and was downloaded from [Stanford's Digital Repository](https://purl.stanford.edu/xw602fs2985).  

```{r}
# Convert name to island
ci_poly <- ci_shp %>% 
  rename(island = NAME)

# Graph to check - looks pretty good 
ggplot() + 
  geom_sf(data = ca_counties, fill = NA) + 
  geom_sf(data = ci_poly, aes(fill = NAME, color=NAME)) + 
  theme_classic()
```

```{r save}
# Save the islands shapefile to project folder
st_write(ci_poly, file.path(project_data_path, "processed", "spatial", "channel_islands.shp"))

# Convert geometry to wkt
ci_bq <- ci_polys %>% 
  mutate(boundary = st_as_text(geometry)) %>% 
  st_set_geometry(NULL)

# Upload to BigQuery
library(bigrquery)

bq_table(project = "emlab-gcp", table = "channel_islands_boundaries", dataset = "boi_dive_project") %>%
  bq_table_upload(values = ci_bq, fields = ci_bq)
```

## MPAs

```{r}
# Channel Islands MPAs
ci_mpas <- mpa_atlas %>% 
  filter(country == 'USA' & 
           name %in% c('Channel Islands', 
                       'Anacapa Island',
                       'Footprint',
                       'Footprint (Anacapa Channel)',
                       'Scorpion',
                       'Scorpion (Santa Cruz Island)',
                       'Gull Island',
                       'Gull Island (Santa Cruz Island)',
                       'Carrington Point (Santa Rosa Island)',
                       'Skunk Point (Santa Rosa Island)',
                       'South Point',
                       'South Point (Santa Rosa Island)',
                       'Richardson Rock',
                       'Richardson Rock (San Miguel Island)',
                       'Judith Rock (San Miguel Island)',
                       'Harris Point',
                       'Harris Point (San Miguel Island)',
                       'Begg Rock (San Nicolas Island)',
                       'Arrow Point to Lion Head Point (Catalina Island)',
                       'Blue Cavern (Catalina Island) Offshore',
                       'Blue Cavern (Catalina Island) Onshore',
                       'Casino Point (Catalina Island)',
                       'Cat Harbor (Catalina Island)',
                       'Farnsworth Offshore (Catalina Island)',
                       'Farnsworth Onshore (Catalina Island)',
                       "Lover's Cove (Catalina Island)",
                       'Long Point (Catalina Island)',
                       'Santa Barbara Island',
                       'Painted Cave (Santa Cruz Island)')) %>% 
  dplyr::mutate(gov_type = ifelse(is.na(gov_type), "Federal", gov_type)) %>% 
  dplyr::select(mpa_id, name, mpa_type = designatio, status, status_yr = status_yea, 
                gov_type, no_take, fishing, area_km2 = calc_m_are) %>% 
  # Remove the "special closure" area, "fish habitat area" 
  filter(!mpa_id %in% c(7606, 8670))

# CI polygons
ci_polys <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands.shp")) %>% 
  filter(island != 'San Clemente')

# Plot to check 
ci_fed <- ci_mpas %>% 
  filter(gov_type == 'Federal' & mpa_id != 8688) 

ci_nms <- ci_mpas %>% 
  filter(mpa_id == 8688)

ci_state <- ci_mpas %>% 
  filter(gov_type == 'State')

mpa_map <- ggplot() + 
  geom_sf(data = ci_polys, fill = NA, color = 'black') + 
  geom_sf(data = ci_nms, fill = NA, aes(color = mpa_type)) + 
  geom_sf(data = ci_fed, aes(fill = mpa_type), alpha = 0.4) + 
  geom_sf(data = ci_state, aes(fill = mpa_type), alpha = 0.6) + 
  scale_fill_manual(values = c('slategray3', 'tomato', 'darkcyan', 'firebrick')) +
  scale_color_manual(values = c('midnightblue')) + 
  labs(color = "",
       fill = 'Type') + 
  theme_bw() + 
  # move color legend under the fill legend 
  guides(color = guide_legend(order = 2),
         fill = guide_legend(order = 1, nrow = 2)) +
  theme(legend.margin=margin(0,0,0,0, unit="cm"),
        legend.position = 'bottom')
```

```{r refine-mpas, message=F, warning=F}
# Revise the no take designation for Fed. Conservation Area (some fishing allowed)
# Add column for island 
ci_mpas_labeled <- ci_mpas %>% 
  mutate(fishing = ifelse(mpa_type == 'Federal Marine Conservation Area', 'Some Restrictions', fishing)) %>% 
  mutate(no_take = ifelse(fishing == "No", "All", "Part")) %>%
  mutate(island = case_when(str_detect(name, "Anacapa") ~ "Anacapa",
                            str_detect(name, "Catalina") ~ "Catalina",
                            str_detect(name, "Santa Rosa") ~ "Santa Rosa",
                            str_detect(name, "Santa Cruz") ~ "Santa Cruz",
                            str_detect(name, "San Miguel") ~ "San Miguel",
                            str_detect(name, "Santa Barbara") ~ "Santa Barbara",
                            str_detect(name, "San Nicolas") ~ "San Nicolas",
                            name %in% c("Richardson Rock", "Harris Point") ~ "San Miguel",
                            name %in% c("Scorpion", "Gull Island") ~ "Santa Cruz",
                            name == "Channel Islands" ~ "Channel Islands",
                            name == "Footprint" ~ "Anacapa",
                            name == "South Point" ~ "Santa Rosa"))

# Save North CI only
north_ci_mpas <- ci_mpas_labeled %>% 
  filter(island %in% c("San Miguel", "Santa Rosa", "Santa Cruz", "Anacapa"))
```


```{r save-mpas}
# Save the mpa shapefile to project folder
st_write(ci_mpas_labeled, file.path(project_data_path, "processed", "spatial", "channel_islands_mpas.shp"))

# Save north ci mpas
st_write(north_ci_mpas, file.path(project_data_path, "processed", "spatial", "north_channel_islands_mpas.shp"))

# Save map 
ggsave(plot = mpa_map, filename = file.path(project_figure_path, "mpa_map.png"))
```

## Anchorages

Get locations of major ports/harbors for the Channel Islands, and in Santa Barbara, Ventura, and Los Angeles counties. Use GFW anchorages data to try and find the main ports.  

Main ports by Island: 
 - San Miguel: Cuyler Harbor  
 - Santa Rosa: Bechers Bay  
 - Santa Cruz: Prisoners Harbor and Scorpion  
 - Anacapa: East End Landing Cove  
 - Santa Barbara: Landing Cove  
 - Catalina: Two Harbors and Avalon  
 - San Clemente: Wilson Cove and Pyramid Cove  
 
Mainland ports:   
  - Santa Barbara harbor   
  - Ventura harbor   
  - Port Hueneme   
  - Port of Long Beach   
  - Port of Los Angeles   
<br>   

```{r identify-harbors}
# Island ports
# These are just for plotting - get coordinates from google maps 
island_anchorages <- data.frame(harbor = c('Cuyler Harbor', 'Bechers Bay', 'Prisoners Harbor', 'Scorpion', 'Landing Cove',
                                           'Landing Cove', 'Two Harbors', 'Avalon', 'Wilson Cove', 'Pyramid Cove'),
                                island = c('San Miguel', 'Santa Rosa', 'Santa Cruz', 'Santa Cruz', 'Anacapa',
                                           'Santa Barbara', 'Catalina', 'Catalina', 'San Clemente', 'San Clemente'),
                                lat = c(34.051, 34.008, 34.023, 34.048, 34.017,
                                        33.482, 33.442, 33.346, 33.006, 32.823),
                                lon = c(-120.351, -120.047, -119.684, -119.553, -119.362,
                                        -119.030, -118.498, -118.323, -118.557, -118.385))

# Mainland ports
mainland_anchorages <- data.frame(harbor = c("Santa Barbara Harbor", "Ventura Harbor", "Port Hueneme",
                                             "Port of Los Angeles", "Port of Long Beach"),
                                  lat = c(34.406, 34.244, 34.144,
                                          33.708, 33.724),
                                  lon = c(-119.686, -119.270, -119.213,
                                          -118.265, -118.211))

# Combine ports/harbors
combined_harbors <- island_anchorages %>% 
  bind_rows(mainland_anchorages)

# Check on map 
ggplot() + 
  geom_sf(data = ca_counties, fill = NA) + 
  geom_sf(data = ci_shp, fill = NA) + 
  geom_point(data = combined_harbors, aes(x=lon, y=lat, color = harbor)) +
  labs(x = "",
       y = "",
       color = '') + 
  theme_bw() +
  theme(axis.text = element_blank(),
        axis.ticks = element_blank()) #Looks pretty good

# Save anchorages
write_csv(combined_harbors, file.path(project_data_path, "processed", "spatial", "harbors.csv"))
```
