==> src/data-prep/01_vessel_list.Rmd <==
---
title: "Vessel List"
output: html_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE
)
```

```{r libraries, include=F}
# Load libraries
library(tidyverse)
library(janitor)
library(bigrquery)
library(DBI)
library(readxl)
library(sf)

# Load source
source(file.path(here::here(), "src", "common.R"))

# Raw vessel list
all_vessels <- read_xlsx(path = file.path(project_data_path, "raw", "CAdiveboatlist_names.xlsx"),
                         sheet = "Sheet1")

```

# Vessel Names

Create a list of dive vessel names in southern California 

```{r vessel-names}
vessel_names <- all_vessels %>% 
  clean_names() %>% 
  filter(boat != '?') %>% 
  dplyr::select(boat, ca_area, operator) %>% 
  arrange(ca_area) # 42 vessels to look for 

# Upload to BigQuery
bq_table(project = "emlab-gcp", table = "raw_vessel_names", dataset = "boi_dive_project") %>%
  bq_table_upload(values = vessel_names, fields = vessel_names,
                  write_disposition = 'WRITE_TRUNCATE')
```


# Possible Vessel Identities 

Search for these 42 vessels in GFW and try to match vessel identities 

```{r sql-identities}
sql_identities <- "#StandardSQL

WITH 

  ######################
  # List of normalized vessel names 
  
  vessel_names AS(
  SELECT
    boat AS shipname,
    `world-fishing-827.udfs.normalize_shipname`(boat) AS n_shipname
  FROM
    `emlab-gcp.boi_dive_project.raw_vessel_names`
  ),

  ######################
  # Matches from the vessel info tables
  
  vessel_info AS (
    SELECT
      *
    FROM
      `world-fishing-827.gfw_research.vi_ssvid_v20220501` a
    INNER JOIN 
      vessel_names b
    ON(a.registry_info.best_known_shipname = b.n_shipname)
  )
  
  ######################
  # Matches flagged to USA
  # Plus Sunfish match identified through vessel registry 
  
  SELECT 
    * 
  FROM 
    vessel_info
  WHERE best.best_flag = 'USA'
  UNION ALL (
    SELECT
      *,
      'Sunfish' AS shipname,
      'SUNFISH' AS n_shipname
    FROM
      `world-fishing-827.gfw_research.vi_ssvid_v20220501`
    WHERE
      ssvid = '338350134'
  )"
```

```{r run-save-identities}
# Run query and write results to BigQuery
bq_project_query("emlab-gcp",
                 sql_identities,
                 destination_table = bq_table(project = "emlab-gcp",
                                              table = "initial_vessel_identity_matches",
                                              dataset = "boi_dive_project"),
                 use_legacy_sql = FALSE, allowLargeResults = TRUE,
                 write_disposition = 'WRITE_TRUNCATE')
```

```{r all-possible-matches}
sql_matches <- "#StandardSQL

SELECT
  ssvid,
  shipname,
  n_shipname
FROM
  `emlab-gcp.boi_dive_project.initial_vessel_identity_matches`"

possible_matches <- bq_project_query("emlab-gcp", sql_matches) %>% 
  bq_table_download(n_max = Inf)

# Arrange by vessel name and mmsi
possible_matches <- possible_matches %>% 
  arrange(n_shipname, ssvid) #23 vessels with potential matches 
```

# New Boats 

Some vessels already have confirmed mmsi based on the list Emma sent recently. 

```{r add-new-boats}
# New List
new_vessels <- read_xlsx(file.path(project_data_path, "raw", "dive_boat_list.xlsx"),
                         sheet = "Sheet1") %>% 
  row_to_names(1) %>% 
  clean_names() %>% 
  mutate(mmsi = as.numeric(mmsi),
         n_shipname = gsub(" ", "", toupper(ship_name))) %>% 
  mutate(n_shipname = ifelse(n_shipname == 'MVHORIZON', 'HORIZON', n_shipname)) %>% 
  filter(!is.na(mmsi)) %>% 
  dplyr::select(ssvid = mmsi, n_shipname, ship_name)

# Remove these vessels from the possible matches list since these are definitely matched 
vessels_to_research <- possible_matches %>% 
  filter(!n_shipname %in% new_vessels$n_shipname)
```

Searched the potential matches on Marine Traffic to identify which ones are likely our dive vessels based on vessel class and area of operation. This list includes 15 unique vessel names. The following vessels are found in Marine Traffic with a possibly correct MMSI:  

  - Islander: 366813530 (Ventura) or 367036830 (San Diego)  
  - Pacific Moon: 368200240  
  - Ranger: 368926530 (Oxnard but Ranger vessel is supposed to be from San Pedro; gear type 'high speed craft')
  - Sunfish: 338350134  

The MMSI of the remaining vessels were not determined to be correct based on vessel type and operating area:  

  - Asante   
  - Explorer  
  - Fortuna  
  - Great Escape 
  - Great White  
  - Humboldt  
  - Magic  
  - Mr.C  
  - Sand Dollar  
  - Sea Bass

```{r matched-vessels}
# Based on marine traffic findings keep the correct MMSI for each vessel 
matched_marine_traffic <- vessels_to_research %>% 
  filter(ssvid %in% c('338350134', '366813530', '367036830', '368926530')) %>% 
  mutate(ssvid = as.double(ssvid))

# Add to other matched vessels 
matched_vessels <- new_vessels %>% 
  rename(shipname = ship_name) %>% 
  bind_rows(matched_marine_traffic) %>% 
  # Add one match by hand that wasn't in the vessel list 
  bind_rows(data.frame(ssvid = 368200240,
                       n_shipname = 'PACIFICMOON',
                       shipname = 'Pacific Moon'))
```

# Check Tracks

Pull tracks for October 2020 for the 15 matched vessels and check to see if this lines up with activity in the Channel Islands. In particular look at the two MMSI for Islander and the potential MMSI for the Ranger. 

```{r sql-tracks}
sql_tracks <- "#StandardSQL

WITH 

  ######################
  # Good segements 
  
  good_segments AS (
    SELECT
      seg_id
    FROM
      `world-fishing-827.pipe_production_v20201001.research_segs`
    WHERE good_seg
    AND NOT overlapping_and_short
  ),

  ######################
  # Tracks for October 2020 
  
  tracks AS (
    SELECT
      *
    FROM
      `world-fishing-827.pipe_production_v20201001.research_messages`
    WHERE ssvid IN ('367315630',
      '367130890',
      '367166830',
      '367041160',
      '367384840',
      '367544000',
      '338350134', 
      '367616930',
      '367655170',
      '367098480',
      '338776000',
      '368024590', 
      '366813530',
      '367036830',
      '368926530',
      '368200240'
    )
    AND seg_id IN (
      SELECT
        seg_id
      FROM good_segments)
    AND _PARTITIONTIME BETWEEN '2021-10-01' AND '2021-10-31'
  )
  
  ######################
  SELECT * FROM tracks"
```

```{r run-download-tracks}
# Run and download tracks 
tracks <- bq_project_query("emlab-gcp", sql_tracks) %>% 
  bq_table_download(n_max = Inf)
```

```{r map-tracks}
# Central CA coastline
ca_counties <- tigris::counties() %>% 
  filter(GEOID %in% c('06083', '06111', '06037'))

# CI polygons
ci_polys <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands.shp"))

# ssvid list (no tracks for Conception)
ssvids <- unique(tracks$ssvid) 
ssvids <- c('366813530', '367036830','368926530', '368200240') # Priority ssvids to check

# Quick map of California for each of the 13 vessels 
for(i in ssvids){
  
  vessel_id <- ssvids[3]
  
  track <- tracks %>% 
    filter(ssvid == vessel_id) %>% 
    arrange(timestamp)
  
  track_map <- ggplot() + 
    #geom_sf(data = ca_counties) + 
    geom_sf(data = ci_polys) + 
    geom_point(data = track,
               aes(x=lon, y=lat)) +
    labs(title = vessel_id)

  print(track_map)
} #ssvid: 367544000 is not the right one for Explorer; 366813530 or 367036830 could be right one for islander; 368926530 for Ranger doesn't seem right; 368200240 for Pacific Moon maybe right but didn't go anywhere in this track timeframe
```

## AIS Identity for Matched Vessels

The Explorer ssvid isn't correct since that vessel mainly operates in the Gulf of Mexico. There's a chance the Ranger id is correct but I don't think it is, it's listed as a high speed craft in Marine Traffic and the last location was in Hawaii so even though it has some tracks around the Channel Islands it doesn't seem like it's stopping at the islands regularly. Both of the ssvids for Islander might be correct as they are both showing tracks around the islands; 366813530 more around Santa Cruz and Anacapa and 367036830 more around San Clemente and Catalina.  

```{r sql-vessel-list}
sql_ais_list <- "#StandardSQL

SELECT
  ssvid,
  n_shipname,
  shipname,
  registry_info.best_known_callsign AS callsign,
  EXTRACT(DATE FROM activity.first_timestamp) AS ais_first_date,
  EXTRACT(DATE FROM activity.last_timestamp) AS ais_last_date
FROM
  `emlab-gcp.boi_dive_project.initial_vessel_identity_matches`
WHERE ssvid IN ('338350134',
  '338776000',
  '366813530',
  '367036830',
  '367041160',
  '367098480',
  '367130890', 
  '367166830',
  '367315630',
  '367384840',
  '367616930',
  '367655170',
  '368024590',
  '368926530')
UNION ALL (
    SELECT
      ssvid,
      ais_identity.n_shipname_mostcommon.value AS n_shipname,
      'Pacific Moon' AS shipname,
      registry_info.best_known_callsign AS callsign,
      EXTRACT(DATE FROM activity.first_timestamp) AS ais_first_date,
      EXTRACT(DATE FROM activity.last_timestamp) AS ais_last_date
    FROM
      `world-fishing-827.gfw_research.vi_ssvid_v20220501`
    WHERE
      ssvid = '368200240'
  )"
```

```{r run-download-vessel-list}
# Run query and download results
ais_list <- bq_project_query("emlab-gcp", sql_ais_list) %>% 
  bq_table_download(n_max = Inf)
```

## Final List

The final list of dive vessels includes the vessel MMSI, shipname, IMO, California area (e.g. Santa Barbara, Ventura, Long Beach), operator, first date active in AIS, and last date active in AIS (NA if still currently active). This list has 15 vessels of the 25 (60%) that are likely still operational in southern California.  

```{r final-ais-list}
# Current AIS list + operator info 
current_vessels <- ais_list %>% 
  left_join(vessel_names, 
            by = c('shipname' = 'boat'))

# Add AIS list table to BQ 
bq_table(project = "emlab-gcp", table = "ais_vessel_list", dataset = "boi_dive_project") %>%
  bq_table_upload(values = current_vessels, fields = current_vessels,
                  write_disposition = 'WRITE_TRUNCATE')

# Save csv
write_csv(current_vessels, file.path(project_data_path, "processed", "ais_vessel_list.csv"))
```
