==> src/data-prep/03b_mpa_dive_site_overlap.Rmd <==
---
title: 'Dive Site MPA Overlap'
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE
)
```

```{r libraries, include=F}
# Load libraries
library(tidyverse)
library(sf)
library(bigrquery)
library(lubridate)
library(paletteer)

# Load source
source(file.path(here::here(), "src", "common.R"))

# Dive sites
dive_sites <- read_csv(file.path(project_data_path, "processed", "dive_sites_1hr30min_1vessel_labeled.csv"))
```

```{r mapping, include=F}
# Channel Islands shp
ci_shp <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands.shp"))

# MPA Boundaries 
mpa_shp <- st_read(file.path(project_data_path, "processed", "spatial", "north_channel_islands_mpas.shp"))
```

## Overview 

In this document we look at the overlap of dive sites and MPAs. We are counting a dive site as any grid cell with at least 1 visits, where a visit is defined as a dive boat that has remained stationary in that grid cell for at least 1.5 hours on a given day. We classify each dive site as being within an MPA, outside an MPA, or within the buffer (e.g., within 500 meters of an MPA boundary). For each dive site we add the following information: 

  - site_type: dive site type (in_mpa, in_buffer, outside_mpa)  
  - mpa_name: name of MPA, `NA` if site is outside MPA   
  - mpa_type: type or designation of MPA (Federal Marine Reserve, Federal Marine Conservation Area, State Marine Reserve, State Marine Conservation Area), `NA` if outside MPA    
  - no_take: the level of no take in the area (full for Federal/State Marine Reserves; part for Federal/State Marine Conservation Areas; none if outside MPA)  

## Buffer MPAs

Map of Channel Islands MPAs based on the 2020 MPA Atlas. For determining if dive sites are within an MPA, we look at MPAs with the following designations:  

  - State Marine Reserves   
  - State Marine Conservation Areas  
  - Federal Marine Reserve  
  - Federal Marine Conservation Area  

```{r mpa-map}
# Remove the "National Marine Sanctuary" boundary from shp
mpas <- mpa_shp %>% 
  filter(mpa_id != 8688)

ggplot() + 
  geom_sf(data = ci_shp, fill = NA, color = 'black') + 
  geom_sf(data = mpas, aes(fill = mpa_type), alpha = 0.6) + 
  scale_fill_manual(values = c('slategray3', 'tomato', 'darkcyan', 'firebrick')) + 
  labs(fill = "MPA Type",
       x="",
       y="") + 
  theme_bw()

## Join Federal and State Reserves/Conservation Areas for secondary analysis 
# Update names for adjacent MPAs
mpa_join <- mpas %>% 
  mutate(join_name = str_remove(name, " \\(.*")) %>% 
  mutate(join_name = str_remove(join_name, "Offshore"),
         join_name = str_remove(join_name, "Onshore")) %>% 
  mutate(mpa_type = case_when(join_name == 'Anacapa Island' ~ "Federal/State Reserve & Conservation Area",
                              join_name %in% c('Blue Cavern', 'Farnsworth ') ~ "State Marine Conservation Area",
                              join_name %in% c('Footprint', 'Gull Island', 'Harris Point', 'Richardson Rock',
                                               'Santa Barbara Island', 'Scorpion', 'South Point') ~ "Federal/State Marine Reserve",
                              TRUE ~ mpa_type)) 

mpa_join <- mpa_join %>% group_by(join_name, mpa_type) %>% summarize()

# check
ggplot() + 
  geom_sf(data = ci_shp, fill = NA, color = 'black') + 
  geom_sf(data = mpa_join, aes(fill = join_name), alpha = 0.6) + 
  #scale_fill_manual(values = c('slategray3', 'tomato', 'darkcyan', 'firebrick')) + 
  labs(fill = "",
       x="",
       y="") + 
  theme_bw() # Looks right but some weird lines still in the polygons? 
```

We create two types of MPA buffers. For one we consider a 500 meter buffer around only Federal/State Marine Reserves, for the second we create a 500 meter buffer around the grouped Federal/State Marine Reserves and the Federal/State Conservation Areas.  

We add two additional buffer lengths of 1 km (1000 meters) and 1.5 km (1500 meters) to test for differences in dive behavior in the buffers, especially for night dives during lobster season.

```{r buffer-mpas, include=F}
## Buffer only reserves - use NAD83 for distance 
reserves_nad83 <- mpas %>% 
  filter(mpa_type %in% c("Federal Marine Reserve", "State Marine Reserve")) %>% 
  st_transform(crs = 4269)

# Create 500 meter buffer around all MPA boundaries
reserve_buff_500m <- st_buffer(reserves_nad83, dist=500)

# Remove the polygon to just get the buffer
reserve_buff_boundary_500m  <- rmapshaper::ms_erase(reserve_buff_500m, reserves_nad83)

## Buffer combined Reserves and Conservation Areas
mpa_join_nad83 <- st_transform(mpa_join, crs=4269)
joint_buff_500m <- st_buffer(mpa_join_nad83, dist=500)

# Remove polygon to just get buffer
joint_buff_boundary_500m <- rmapshaper::ms_erase(joint_buff_500m, mpa_join_nad83)

## 1000 m buffer
# Buffer Marine Reserves 
reserve_buff_1000m <- st_buffer(reserves_nad83, dist=1000)
reserve_buff_boundary_1000m  <- rmapshaper::ms_erase(reserve_buff_1000m, reserves_nad83)

## Buffer combined Reserves and Conservation Areas
joint_buff_1000m <- st_buffer(mpa_join_nad83, dist=1000)
joint_buff_boundary_1000m <- rmapshaper::ms_erase(joint_buff_1000m, mpa_join_nad83)

## 1500 m buffer
# Buffer Marine Reserves
reserve_buff_1500m <- st_buffer(reserves_nad83, dist=1500)
reserve_buff_boundary_1500m  <- rmapshaper::ms_erase(reserve_buff_1500m, reserves_nad83)

## Buffer combined Reserves and Conservation Areas
joint_buff_1500m <- st_buffer(mpa_join_nad83, dist=1500)
joint_buff_boundary_1500m <- rmapshaper::ms_erase(joint_buff_1500m, mpa_join_nad83)


## Save 500 meter buffer 
reserve_buff_save <- reserve_buff_boundary_500m %>% 
  st_transform(crs = 4326)

joint_buff_save <- joint_buff_boundary_500m %>% 
  st_transform(crs = 4326)

st_write(reserve_buff_save, file.path(project_data_path, "processed", "spatial", "reserve_buffer_500m.shp"))
st_write(joint_buff_save, file.path(project_data_path, "processed", "spatial", "joint_buffer_500m.shp"))
```

## Classify Dive Sites

We classify each of the 2,605 dive sites as being within a MPA, within a buffer, or outside a MPA.   
First we do this considering "in MPA" as being only within a Federal or State Marine Reserve. We repeat the process for each of the 3 buffer distances.    

```{r classify-dive-sites-reserves, echo=F}
# Crs of mpa shp 
target_crs <- st_crs(reserves_nad83)

# Create spatial points from dive sites
dive_sites_sf <- st_as_sf(dive_sites,
                          coords = c('lon_bin', 'lat_bin'),
                          crs = target_crs)

# Reserves only 
sites_in_reserve <- st_intersection(dive_sites_sf, reserves_nad83) #278 unique sites 

# Dive sites within the buffer 
## 500 meters
sites_in_reserve_buff_500 <- st_intersection(dive_sites_sf, reserve_buff_boundary_500m) #160 unique sites

## 1000 meters
sites_in_reserve_buff_1000 <- st_intersection(dive_sites_sf, reserve_buff_boundary_1000m) #322 unique sites

## 1500 meters 
sites_in_reserve_buff_1500 <- st_intersection(dive_sites_sf, reserve_buff_boundary_1500m) #478 unique sites

# Combine 
dive_site_reserve_500 <- sites_in_reserve %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_reserve_buff_500 %>% 
              mutate(site_category = "in_buffer"))

dive_site_reserve_1000 <- sites_in_reserve %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_reserve_buff_1000 %>% 
              mutate(site_category = "in_buffer"))

dive_site_reserve_1500 <- sites_in_reserve %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_reserve_buff_1500 %>% 
              mutate(site_category = "in_buffer"))

# What sites aren't in MPAs or the buffer?
ids_500 <- dive_site_reserve_500$site_id
ids_1000 <- dive_site_reserve_1000$site_id
ids_1500 <- dive_site_reserve_1500$site_id

missing_ids_500 <- setdiff(dive_sites$site_id, ids_500) #2,090 outside 
missing_ids_1000 <- setdiff(dive_sites$site_id, ids_1000) #1,928 outside 
missing_ids_1500 <- setdiff(dive_sites$site_id, ids_1500) # 1,772 outside 

outside_reserve_500 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_500) %>% 
  mutate(site_category = 'outside_mpa')

outside_reserve_1000 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_1000) %>% 
  mutate(site_category = 'outside_mpa')

outside_reserve_1500 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_1500) %>% 
  mutate(site_category = 'outside_mpa')

# Add missing to the list 
all_site_reserve_500 <- dive_site_reserve_500 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, mpa_id, name, mpa_type) %>% 
  bind_rows(outside_reserve_500) %>% 
  arrange(site_id)

all_site_reserve_1000 <- dive_site_reserve_1000 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, mpa_id, name, mpa_type) %>% 
  bind_rows(outside_reserve_1000) %>% 
  arrange(site_id)

all_site_reserve_1500 <- dive_site_reserve_1500 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, mpa_id, name, mpa_type) %>% 
  bind_rows(outside_reserve_1500) %>% 
  arrange(site_id)

## Save dive sites by category
reserve_categories_save_500 <- all_site_reserve_500 %>% st_set_geometry(NULL)
reserve_categories_save_1000 <- all_site_reserve_1000 %>% st_set_geometry(NULL)
reserve_categories_save_1500 <- all_site_reserve_1500 %>% st_set_geometry(NULL)


write_csv(reserve_categories_save_500, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_reserves_500m.csv"))
write_csv(reserve_categories_save_1000, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_reserves_1000m.csv"))
write_csv(reserve_categories_save_1500, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_reserves_1500m.csv"))
```
<br>  

We then look at "in MPA" as being in either Federal/State Marine Reserves or Federal/State Conservation Areas.  

```{r classify-dive-sites-joint, echo=F}
## Joint reserves and conservation areas 
# Dive sites within the MPA boundary 
sites_in_joint <- st_intersection(dive_sites_sf, mpa_join_nad83) #459 unique sites 

# Dive sites within the buffer 
sites_in_joint_buff_500 <- st_intersection(dive_sites_sf, joint_buff_boundary_500m) #283 unique sites
sites_in_joint_buff_1000 <- st_intersection(dive_sites_sf, joint_buff_boundary_1000m) #636 unique sites
sites_in_joint_buff_1500 <- st_intersection(dive_sites_sf, joint_buff_boundary_1500m) #845 unique sites

# Combine 
dive_site_joint_500 <- sites_in_joint %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_joint_buff_500 %>% 
              mutate(site_category = "in_buffer"))

dive_site_joint_1000 <- sites_in_joint %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_joint_buff_1000 %>% 
              mutate(site_category = "in_buffer"))

dive_site_joint_1500 <- sites_in_joint %>% 
  mutate(site_category = "in_mpa") %>% 
  bind_rows(sites_in_joint_buff_1500 %>% 
              mutate(site_category = "in_buffer"))

# What sites aren't in MPAs or the buffer?
ids_500 <- dive_site_joint_500$site_id
ids_1000 <- dive_site_joint_1000$site_id
ids_1500 <- dive_site_joint_1500$site_id

missing_ids_500 <- setdiff(dive_sites$site_id, ids_500) #1,793 outside 
missing_ids_1000 <- setdiff(dive_sites$site_id, ids_1000) #1,572 outside 
missing_ids_1500 <- setdiff(dive_sites$site_id, ids_1500) #1,413 outside 

outside_joint_500 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_500) %>% 
  mutate(site_category = 'outside_mpa')

outside_joint_1000 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_1000) %>% 
  mutate(site_category = 'outside_mpa')

outside_joint_1500 <- dive_sites_sf %>% 
  filter(site_id %in% missing_ids_1500) %>% 
  mutate(site_category = 'outside_mpa')

# Add missing to the list 
all_site_joint_500 <- dive_site_joint_500 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, name = join_name, mpa_type) %>% 
  bind_rows(outside_joint_500) %>% 
  arrange(site_id) %>% 
  # 6 sites within buffer of two Catalina State Conservation Areas
  # Combine names to keep correct number of sites 
  mutate(name = ifelse(site_id %in% c(1808:1814), "Casino Point/Lover's Cove", name)) %>% 
  distinct()

all_site_joint_1000 <- dive_site_joint_1000 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, name = join_name, mpa_type) %>% 
  bind_rows(outside_joint_1000) %>% 
  arrange(site_id) %>% 
  # Fix any duplicated buffer/mpa 
  mutate(name = case_when(site_id %in% c(1808:1831) ~ "Casino Point/Lover's Cove", 
                          site_id %in% c(1972:1974, 1982:1984, 1987:1993, 
                                         1996:2008, 2010:2019, 2025:2027, 2033) ~ "Blue Cavern/Arrow Point to Lion Head Point",
                          TRUE ~ name),
         site_category = ifelse(site_id %in% c(1955:1971, 1977:1979, 1995, 2009, 
                                               2020:2024, 2029:2032, 2034:2036, 2038:2054, 
                                               2060:2061, 2074:2082, 2085:2090, 2094, 2101:2105, 
                                               2110:2111, 2114:2115, 2125, 2143:2144), "in_mpa", site_category)) %>% 
  distinct()

all_site_joint_1500 <- dive_site_joint_1500 %>% 
  dplyr::select(site_id, avg_distance_shore_m, nearest_island, island_group, site_category, name = join_name, mpa_type) %>% 
  bind_rows(outside_joint_1500) %>% 
  arrange(site_id) %>% 
  # Fix any duplicated buffer/mpa 
  mutate(name = case_when(site_id %in% c(1808:1831, 1833:1841) ~ "Casino Point/Lover's Cove", 
                          site_id %in% c(1994, 
                                         1996:2008, 2010:2019, 2025:2027, 2033, 2066:2068, 2071:2073) ~ "Blue Cavern/Arrow Point to Lion Head Point",
                          site_id %in% c(1963:1964, 1967, 1972:1976, 1982:1984, 1987:1993) ~ "Blue Cavern/Arrow Point to Lion Head Point/Cat Harbor",
                          site_id %in% c(2120:2123, 2133:2135, 2155) ~ "Arrow Point to Lion Head Point",
                          TRUE ~ name),
         mpa_type = ifelse(site_id %in% c(2120:2123, 2133:2135, 2155), "State Marine Conservation Area", mpa_type),
         site_category = case_when(site_id %in% c(1922:1935, 1944, 1955:1962, 1965:1966, 1968:1971, 1977:1979,  
                                                  1995, 2009, 2020:2024, 2029:2032, 2034:2036, 2038:2054, 
                                                  2060:2061, 2074:2075, 2077:2082, 2085:2091, 2094, 2101:2105, 
                                                  2110:2111, 2114:2115, 2125, 2143:2144) ~ "in_mpa", 
                                   site_id %in% c(2120:2123, 2133:2135, 2155) ~ "in_buffer",
                                   TRUE ~ site_category)) %>% 
  distinct()

## Save dive sites by category
joint_categories_save_500 <- all_site_joint_500 %>% st_set_geometry(NULL)
joint_categories_save_1000 <- all_site_joint_1000 %>% st_set_geometry(NULL)
joint_categories_save_1500 <- all_site_joint_1500 %>% st_set_geometry(NULL)

write_csv(joint_categories_save_500, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_joint_500m.csv"))
write_csv(joint_categories_save_1000, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_joint_1000m.csv"))
write_csv(joint_categories_save_1500, file.path(project_data_path, "processed", "dive_sites_by_mpa_category_joint_1500m.csv"))
```
